starting the fix
basic logic:
For each item slot
Check if your new item can go in that slot. If it can:
Check what the amount of items in that slot, plus the amount of items produced
Minus the Itemstack.getMaxStack is.
If that’s less than 0, insert.
If itks less than number you have spare, insert what you can.

If both fail, move on to the next slot and check it.